requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.3.4
PyNaCl>=1.5.0

